Fetcher: better timings and error handling.
The DreamHost pop3 server continues to be incredibly flaky. The extra
time between connections makes it less likely to just give an error on
connect, higher minimum messages makes the fetcher less likely to check
over and over and over (at the cost of a little latency in archiving).

The teardown and setup in the exception handling didn't work; it can't
change the @pop object that's running the each it's inside. So instead,
we just return out and let the caller keep trying if it wants.