ThreadWorker: don't store Messages in multiple Threads
If a stored Thread got a new root Message, it would be stored under the new
Call Number without removing the old Thread. As scary as it is to write a delete
method (Riak does not support renames), this stores under the new Call Number
and then removes the old Thread. So an error results in duplicate data rather
than lost data; even though Threads could simply be recreated from their
Messages if that happened.

This does mean there's a race condition: if a visitor tries to load a Message
at the same time as its Thread is being saved under a new Call Number they'll
get a TooManyFound exception, rather than either thread. If this becomes a
problem in practice it should be OK to remove that check and return either
Thread (though preferably the new one) to the web view. Right now with the
system so new I have little trust built up that I won't dupe Messages in
Threads, so it stays.