ThreadRepo: check for later threads to sort a message with
Debugging this by repeatedly sorting mud-dev/2004/03 and /04. The 'Sun
Middleware' thread is especially interesting as many messages tweak the
subject by dropping out a random space (confusing the subject == and length
comparisons). Seems to be a couple bugs, both in conversation_for? missing
that a conversation should be combined and in threading not picking the
correct root.