Several fixes to threading Messages without references.
Ugh, parent_messages_without_references wasn't learning that it was seeing a
better message; test didn't catch it because its good message was already the
root, so it defaulted through to that. I still need to break that method down
and test it more thoroughly.

Drove out bugs with real-world Thread fixture. I hand-threaded these so I know
they're correct.